add watchedOr and test cases for min
