+++ conjure_oxide/src/rules/cnf.rs
@@ -57,53 +57,3 @@ fn distribute_not_over_or(expr: &Expr, _: &Model) -> ApplicationResult {
         _ => Err(ApplicationError::RuleNotApplicable),
     }
 }
-
-/**
-* Apply the Distributive Law to expressions like `Or([..., And(a, b)])`
-
-* ```text
-* or(and(a, b), c) = and(or(a, c), or(b, c))
-* ```
- */
-#[register_rule(("CNF", 100))]
-fn distribute_or_over_and(expr: &Expr, _: &Model) -> ApplicationResult {
-    fn find_and(exprs: &[Expr]) -> Option<usize> {
-        // ToDo: may be better to move this to some kind of utils module?
-        for (i, e) in exprs.iter().enumerate() {
-            if let Expr::And(_, _) = e {
-                return Some(i);
-            }
-        }
-        None
-    }
-
-    match expr {
-        Expr::Or(_, exprs) => match find_and(exprs) {
-            Some(idx) => {
-                let mut rest = exprs.clone();
-                let and_expr = rest.remove(idx);
-
-                match and_expr {
-                    Expr::And(metadata, and_exprs) => {
-                        let mut new_and_contents = Vec::new();
-
-                        for e in and_exprs {
-                            // ToDo: Cloning everything may be a bit inefficient - discuss
-                            let mut new_or_contents = rest.clone();
-                            new_or_contents.push(e.clone());
-                            new_and_contents.push(Expr::Or(metadata.clone(), new_or_contents))
-                        }
-
-                        Ok(Reduction::pure(Expr::And(
-                            metadata.clone(),
-                            new_and_contents,
-                        )))
-                    }
-                    _ => Err(ApplicationError::RuleNotApplicable),
-                }
-            }
-            None => Err(ApplicationError::RuleNotApplicable),
-        },
-        _ => Err(ApplicationError::RuleNotApplicable),
-    }
-}
