+++ conjure_oxide/tests/integration/basic/min/02/input.essence
@@ -1,0 +1,2 @@
+find a,b : int(1..3)
+such that min([a,b]) <= 2
