+++ conjure_oxide/tests/integration/basic/min/02/input.expected-parse.serialised.json
@@ -1,0 +1,84 @@
+{
+  "constraints": {
+    "Leq": [
+      {
+        "dirtyclean": false
+      },
+      {
+        "Min": [
+          {
+            "dirtyclean": false
+          },
+          [
+            {
+              "Reference": [
+                {
+                  "dirtyclean": false
+                },
+                {
+                  "UserName": "a"
+                }
+              ]
+            },
+            {
+              "Reference": [
+                {
+                  "dirtyclean": false
+                },
+                {
+                  "UserName": "b"
+                }
+              ]
+            }
+          ]
+        ]
+      },
+      {
+        "Constant": [
+          {
+            "dirtyclean": false
+          },
+          {
+            "Int": 2
+          }
+        ]
+      }
+    ]
+  },
+  "variables": [
+    [
+      {
+        "UserName": "a"
+      },
+      {
+        "domain": {
+          "IntDomain": [
+            {
+              "Bounded": [
+                1,
+                3
+              ]
+            }
+          ]
+        }
+      }
+    ],
+    [
+      {
+        "UserName": "b"
+      },
+      {
+        "domain": {
+          "IntDomain": [
+            {
+              "Bounded": [
+                1,
+                3
+              ]
+            }
+          ]
+        }
+      }
+    ]
+  ]
+}
