+++ conjure_oxide/tests/integration/basic/min/02/input.expected-rewrite.serialised.json
@@ -11,22 +11,22 @@
               "dirtyclean": false
             },
             {
-              "Constant": [
-                {
-                  "dirtyclean": false
-                },
-                {
-                  "Int": 3
-                }
-              ]
-            },
-            {
               "Reference": [
                 {
                   "dirtyclean": false
                 },
                 {
                   "MachineName": 2
+                }
+              ]
+            },
+            {
+              "Constant": [
+                {
+                  "dirtyclean": false
+                },
+                {
+                  "Int": 2
                 }
               ]
             },
