+++ conjure_oxide/tests/integration/basic/min/03/input.essence
@@ -1,0 +1,2 @@
+find a : int(1..3), b : int(2..4)
+such that min([a,b]) <= 2
