+++ conjure_oxide/tests/integration/basic/min/03/input.expected-minion.solutions.json
@@ -1,0 +1,37 @@
+[
+  {
+    "2": 1,
+    "a": 1,
+    "b": 2
+  },
+  {
+    "2": 1,
+    "a": 1,
+    "b": 3
+  },
+  {
+    "2": 1,
+    "a": 1,
+    "b": 4
+  },
+  {
+    "2": 2,
+    "a": 2,
+    "b": 2
+  },
+  {
+    "2": 2,
+    "a": 2,
+    "b": 3
+  },
+  {
+    "2": 2,
+    "a": 2,
+    "b": 4
+  },
+  {
+    "2": 2,
+    "a": 3,
+    "b": 2
+  }
+]
