+++ conjure_oxide/tests/integration/basic/min/03/input.expected-rewrite.serialised.json
@@ -11,22 +11,22 @@
               "dirtyclean": false
             },
             {
-              "Constant": [
-                {
-                  "dirtyclean": false
-                },
-                {
-                  "Int": 3
-                }
-              ]
-            },
-            {
               "Reference": [
                 {
                   "dirtyclean": false
                 },
                 {
                   "MachineName": 2
+                }
+              ]
+            },
+            {
+              "Constant": [
+                {
+                  "dirtyclean": false
+                },
+                {
+                  "Int": 2
                 }
               ]
             },
@@ -495,8 +495,8 @@
           "IntDomain": [
             {
               "Bounded": [
-                1,
-                3
+                2,
+                4
               ]
             }
           ]
